backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
httpx==0.27.0
beautifulsoup4==4.12.3
requests==2.31.0
duckduckgo-search==5.2.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.4
python-multipart==0.0.9
aiofiles==23.2.1
jinja2==3.1.3
python-magic==0.4.27
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.21
unstructured[local-inference]>=0.12.5
yt-dlp>=2024.3.10
openai-whisper>=20231117
ffmpeg-python>=0.2.0
edge-tts>=6.1.10
pymupdf>=1.23.26
pydantic-settings>=2.2.1
pynvml>=11.5.0
psutil>=5.9.8
faiss-cpu>=1.8.0
python-docx>=1.1.0
python-pptx>=0.6.23
reportlab>=4.1.0
trafilatura>=1.8.1
jinja2>=3.1.3
aiofiles>=23.2.1
